workspacesView: Use side-by-side layout
Currently, WorkspacesView positions each workspace on a
per-page layout, each page with the allocated width and
height of WorkspaceView. This layout doesn't work well
with horizontal workspaces.

Layout workspaces side by side, instead of per page. The
layout is influenced by a "snap axis", which reflect the
different behaviors exposed in the mockup. This axis
represents which sides of the workspace will be used to
fill the available space. The vertical axis makes the
workspaces fill the available space vertically, event if
it means overflowing horizontally; and vice-versa.

The vertical snap axis is always used for now. Next commits
will make it switch to the horizontal snap axis when in the
app grid state, and when dragging windows.

The translation_{x,y} also needed to reflect the switch to
a side-by-side layout, and use the geometry of the workspaces
to determine the offset. Notice that, when the snap axis is
horizontal, there's no translation applied.