workspacesView: Center active workspace in allocation
When vertically snapping, WorkspacesView currently allocates workspaces
side-by-side, then applies an extra step of translation to center to
the active workspace. This extra step is done in an idle timeout, to
avoid a nasty pick-while-picking case. However, it also means that the
translation can be out of sync with the current allocation state of
workspaces, producing some bad visuals when running transitions.

Move the translation of workspaces to the allocation code itself, and
remove the extra translation step.