overviewControls/controlsManager: Use a custom layout manager
In the future, we want to tightly control the state of the
layout throught gestures, which requires hooking everything
together with adjustments. This is the first step in this
direction.

Here, we add a new custom layout manager for ControlsManager
that allocates the search entry, the view selector, and the
Dash, vertically.

Remove DashSpacer since it is unused now.