viewSelector: Use state adjustment for apps page
Currently, AppsPageContainer reacts to showAppsButton and
transitions between app grid and window picking states on
its own. In the future, we want full control over this.

ControlsManager already has a state adjustment that represents
all possible overview states. Progragate this adjustment up
to AppsPageContainer, and use it to drive the transition.

This requires moving the callback to the showAppsButton to
ControlsManager, since now it control the state adjustment
itself, not AppsPageContainer's adjustment.