workspace: Add cover panel to handle click events
When the workspace is not the active one, we want to switch
to it on click. For that to happen, it needs to cover the
workspace with an actor, to prevent the window preview from
handling the click event itself.

Add a cover actor to Workspace, and a click action that simply
moves to that particular workspace when it's not the active one,
or hide the overview if it's the active one.

Restack the cover actor to be above the wallpaper for the active
workspace, and above all window previews if not the active one.

At last, make sure the click action is released when any the
workspace is allocated. That avoids activating the workspace when
dragging.