overviewControls: Incorporate AppDisplay and WorkspacesDisplay
Move AppDisplay and WorkspacesDisplay from ViewSelector to ControlsManager.
This allows to always allocate the correct size for AppDisplay, and will
enable for a plethora of further improvements. The end goal is to completely
remove ViewSelector, and let ControlsManager handle the layout of everything
that's visible in the overview.

For now, replace the apps page with a dummy actor in ViewSelector.

Adjust various callers around the codebase to not access the ViewSelector
directly from the overview anymore.